requirements.txt
async-timeout==4.0.2
backoff==2.1.2
blinker==1.5
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
docopt==0.6.2
ecs-logging==2.0.0
elastic-apm==6.12.0
Flask==2.2.2
googleapis-common-protos==1.56.4
grpcio==1.48.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.12.0
opentelemetry-exporter-otlp==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-exporter-otlp-proto-http==1.12.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-redis==0.33b0
opentelemetry-instrumentation-wsgi==0.33b0
opentelemetry-proto==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
packaging==21.3
pipreqs==0.4.11
protobuf==3.20.2
psutil==5.9.2
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
six==1.16.0
structlog==22.1.0
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
